requirement.txt
alembic==1.7.3
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.1
h11==0.12.0
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
